requirements.txt
python-dotenv
pandas
numpy
scikit-learn
pymongo[srv]==3.6
pymongo


